Refactor classes into multiple files
